(QA-2038) Rename lib directory
