+++ vmpooler_client/commands/lifetime.py
